quickly implemented fillStateOdBodiesRandomly and did some minor changes here and there

+++ include/Body.hpp
@@ -50,7 +50,7 @@ public:
 
 private:
     int size; //< radius in km
-    int weight; //< weight in tons
+    double weight; //< weight in tons
     Vector2D pos; //<position vector
     Vector2D vel; //<velocity vector
     Vector2D acc; //<acceleration vector
